feat(auth-center): find member profile to verify if the user is authenticated with spring cloud nacos
