+++ .idea/encodings.xml
@@ -1,10 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/knocknut-auth-center/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/knocknut-auth-center/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knocknut-authcenter/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knocknut-authcenter/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knocknut-commons/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knocknut-commons/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/knocknut-database-migration/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/knocknut-database-migration/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knocknut-members/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knocknut-members/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/knoutnut-core/src/main/java" charset="UTF-8" />
