+++ docker-compose.yml
@@ -1,23 +1,35 @@
-version: "3.7"
+version: "3.8"
 
 services:
+  nacos:
+    container_name: knocknut-nacos
+    image: zhusaidong/nacos-server-m1:2.0.3
+    restart: always
+    env_file:
+      - ./docker/env/nacos-standalone-mysql.env
+    ports:
+      - "8848:8848"
+      - "9848:9848"
+      - "9849:9849"
+    depends_on:
+      mysql:
+        condition: service_healthy
+
   mysql:
     container_name: knocknut-mysql
     image: mysql:8.0.31
-    restart: always
     ports:
-      - 3306:3306
+      - "3306:3306"
     volumes:
       - knocknut-data:/var/lib/mysql
       - ./sql-initialization:/docker-entrypoint-initdb.d
-    command: [
-      '--character-set-server=utf8mb4',
-      '--collation-server=utf8mb4_general_ci'
-    ]
-    environment:
-      MYSQL_ROOT_PASSWORD: knocknut
-      MYSQL_USER: devclay
-      MYSQL_PASSWORD: clay@knocknut
-      LANG: C.UTF-8
+    env_file:
+      - ./docker/env/mysql.env
+    healthcheck:
+      test: [ "CMD", "mysqladmin", "ping", "-h", "localhost" ]
+      interval: 5s
+      timeout: 10s
+      retries: 10
+
 volumes:
   knocknut-data:
