+++ knocknut-auth-center/src/main/java/dev/claycheng/KnocknutAuthCenterApplication.java
@@ -1,0 +1,14 @@
+package dev.claycheng;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@EnableFeignClients
+@SpringBootApplication
+public class KnocknutAuthCenterApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(KnocknutAuthCenterApplication.class, args);
+  }
+}
