+++ knocknut-auth-center/src/main/java/dev/claycheng/knocknut/domain/User.java
@@ -1,0 +1,19 @@
+package dev.claycheng.knocknut.domain;
+
+import dev.claycheng.api.enums.MemberStatus;
+import java.time.LocalDate;
+import lombok.Data;
+
+@Data
+public class User {
+  private Long id;
+  private Long memberLevelId;
+  private String username;
+  private String password;
+  private String nickname;
+  private String email;
+  private String phone;
+  private MemberStatus status;
+  private String avatar;
+  private LocalDate birthday;
+}
