+++ knocknut-auth-center/src/main/resources/application-standalone.yml
@@ -1,11 +1,8 @@
 spring.flyway:
   enabled: false
-server.port: 55688
+server.port: 5566
 
-spring.cloud:
-  nacos:
-    discovery:
-      server-addr: 127.0.0.1:8848
+
 
 spring.datasource:
   dynamic:
