--- knocknut-authcenter/pom.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <artifactId>knocknut-authcenter</artifactId>
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <artifactId>knocknut</artifactId>
-    <groupId>dev.claycheng</groupId>
-    <version>1.0.0</version>
-  </parent>
-
-  <properties>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-
-</project>
