+++ knocknut-commons/src/main/java/dev/claycheng/common/exception/DataNotFoundException.java
@@ -1,6 +1,8 @@
 package dev.claycheng.common.exception;
 
-public class DataNotFoundException extends Exception {
+import dev.claycheng.common.exception.base.ForbiddenException;
+
+public class DataNotFoundException extends ForbiddenException {
 
   public DataNotFoundException(String message) {
     super(message);
