+++ knocknut-commons/src/main/java/dev/claycheng/common/exception/base/ForbiddenException.java
@@ -1,2 +1,8 @@
-package dev.claycheng.common.exception.base;public class ForbiddenException {
+package dev.claycheng.common.exception.base;
+
+public class ForbiddenException extends RuntimeException {
+
+  public ForbiddenException(String message) {
+    super(message);
+  }
 }
