+++ knocknut-members/src/main/java/dev/claycheng/KnocknutMembersApplication.java
@@ -1,0 +1,12 @@
+package dev.claycheng;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KnocknutMembersApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(KnocknutMembersApplication.class, args);
+  }
+}
