+++ knocknut-members/src/main/java/dev/claycheng/knocknut/domain/enums/Gender.java
@@ -1,0 +1,14 @@
+package dev.claycheng.knocknut.domain.enums;
+
+import dev.claycheng.base.IntEnum;
+
+public enum Gender implements IntEnum {
+  /** */
+  GIRL,
+  BOY,
+  ;
+
+  public Integer getValue() {
+    return ordinal();
+  }
+}
