--- knocknut-members/src/main/java/dev/claycheng/knocknut/members/KnocknutMembersApplication.java
@@ -1,12 +1,0 @@
-package dev.claycheng.knocknut.members;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KnocknutMembersApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(KnocknutMembersApplication.class, args);
-  }
-}
