--- knocknut-members/src/main/java/dev/claycheng/knocknut/members/domain/enums/Gender.java
@@ -1,14 +1,0 @@
-package dev.claycheng.knocknut.members.domain.enums;
-
-import dev.claycheng.base.IntEnum;
-
-public enum Gender implements IntEnum {
-  /** */
-  GIRL,
-  BOY,
-  ;
-
-  public Integer getValue() {
-    return ordinal();
-  }
-}
