+++ knocknut-members/src/main/java/dev/claycheng/knocknut/repository/factory/MemberBuilder.java
@@ -1,0 +1,16 @@
+package dev.claycheng.knocknut.repository.factory;
+
+import dev.claycheng.knocknut.domain.Member;
+import dev.claycheng.knocknut.repository.impl.model.KnMember;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class MemberBuilder {
+  private final MemberMapper memberMapper;
+
+  public Member fromDbEntity(KnMember knMember) {
+    return memberMapper.fromDatabase(knMember);
+  }
+}
