+++ pom.xml
@@ -38,7 +38,7 @@
         <groupId>org.springframework.boot</groupId>
         <scope>import</scope>
         <type>pom</type>
-        <version>2.3.2.RELEASE</version>
+        <version>2.3.12.RELEASE</version>
       </dependency>
       <dependency>
         <artifactId>spring-cloud-dependencies</artifactId>
@@ -67,6 +67,15 @@
         <groupId>com.baomidou</groupId>
         <version>${dynamic-datasource.version}</version>
       </dependency>
+
+      <dependency>
+        <artifactId>spring-cloud-alibaba-dependencies</artifactId>
+        <groupId>com.alibaba.cloud</groupId>
+        <scope>import</scope>
+        <type>pom</type>
+        <version>${spring-cloud-alibaba.version}</version>
+      </dependency>
+
       <dependency>
         <artifactId>mysql-connector-java</artifactId>
         <groupId>mysql</groupId>
@@ -99,7 +108,11 @@
         <groupId>dev.claycheng</groupId>
         <version>1.0.0</version>
       </dependency>
-
+      <dependency>
+        <artifactId>knocknut-database-migration</artifactId>
+        <groupId>dev.claycheng</groupId>
+        <version>1.0.0</version>
+      </dependency>
       <dependency>
         <artifactId>spock-bom</artifactId>
         <groupId>org.spockframework</groupId>
@@ -116,8 +129,9 @@
   <modelVersion>4.0.0</modelVersion>
   <modules>
     <module>knocknut-members</module>
-    <module>knocknut-authcenter</module>
     <module>knocknut-commons</module>
+    <module>knocknut-database-migration</module>
+    <module>knocknut-auth-center</module>
   </modules>
 
   <packaging>pom</packaging>
@@ -133,7 +147,10 @@
     <maven.compiler.target>11</maven.compiler.target>
     <mysql-connector.version>8.0.15</mysql-connector.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <spring-cloud.version>Hoxton.SR9</spring-cloud.version>
+    <spring-cloud-alibaba.version>2.2.7.RELEASE</spring-cloud-alibaba.version>
+    <!-- Nacos: 2.0.3 -->
+    <!-- Ref: https://github.com/alibaba/spring-cloud-alibaba/wiki/%E7%89%88%E6%9C%AC%E8%AF%B4%E6%98%8E       -->
+    <spring-cloud.version>Hoxton.SR12</spring-cloud.version>
     <swagger2.version>2.7.0</swagger2.version>
   </properties>
   <version>1.0.0</version>
